refactor: handle file propagation better
